backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
pinecone-client==3.0.2
sentence-transformers==2.2.2
torch==2.2.0 
openai==1.70.0